+++ controller/kegboard/buzzer.h
@@ -2,13 +2,13 @@
 #define BUZZER_H
 
 struct MelodyNote {
-  int octave;
-  int note;
-  int duration;
+  int8_t octave;
+  int8_t note;
+  int8_t duration;
 };
 
 void setupBuzzer();
-void playMidiNote(int octave, int note);
+void playMidiNote(int8_t octave, int8_t note);
 void playMelody(struct MelodyNote* melody);
 
 #endif // BUZZER_H
